backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "sublease-platform"
version = "0.1.0"
description = "Sub-leasing apartment platform with social features"
license = {text = "GPL-3.0"}
requires-python = ">=3.12"


dependencies = [
    "fastapi[standard]",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.9.1",
    "aiofiles>=24.1.0",
    "pillow>=11.2.1",
]


[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
    "pytest-asyncio>=1.0.0",
]


[tool.ruff]
line-length = 88
target-version = "py312"
exclude = ["alembic/"]

[tool.ruff.lint]
select = ["E", "W", "F","UP","B","SIM","I"]

[tool.pyright]
python-version = "3.12"
typeCheckingMode = "strict"
include=["src"]
exclude = ["alembic/"]




